+++ docker-compose.yaml
@@ -1,4 +1,4 @@
-version: "3.8"
+version: '3.3'
 services:
     service-1:
         container_name: service-1
