+++ service-3/requirements.txt
@@ -1,2 +1,5 @@
 flask
 requests
+flask-SQLAlchemy
+SQLAlchemy
+pymysql
